Code refactoring: Move b-tag weight calculator to helpers
